Cleanup and fixes to prep building in alpine containers
* added typecasts and changed print formats to minimize warnings in tests compile
* changed shebang to look for more standard locations of bash and python
* parametized a few places in runtime/Makefile so alpine container picks up correct gcc location
* a few fixes to C tests to compile cleaner on ALpine

+++ tests/intr_test.c
@@ -52,7 +52,7 @@ void* subrun(void* data)
       }
    }
    printf("**** NOW IS THE END VCPU %d***\n", n);
-   exit(self);
+   exit((uint64_t)self);
 }
 
 TEST intrtest(void)
